# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinancePro API",
    "description": "**FinancePro** is a comprehensive personal finance management platform with AI-powered features for transaction classification, budget forecasting, and intelligent financial insights.\n\n## Features\n- Multi-account management\n- Transaction tracking with AI classification\n- Budget planning and monitoring\n- AI-powered financial insights\n- Advanced analytics and reporting\n\n## Authentication\nAll endpoints (except `/auth/*`) require Bearer JWT token authentication.",
    "termsOfService": "https://financepro.app/terms",
    "contact": {
      "name": "FinancePro Support Team",
      "url": "https://financepro.app/support",
      "email": "support@financepro.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0",
    "x-logo": {
      "url": "https://financepro.app/assets/logo.png",
      "altText": "FinancePro Logo"
    },
    "x-api-id": "financepro-api",
    "x-audience": "external",
    "x-environment": "development"
  },
  "paths": {
...
    "/api/v1/accounts/": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "List all accounts",
        "description": "Retrieve all accounts for the authenticated user, filtered by currently selected financial profiles",
        "operationId": "list_accounts_api_v1_accounts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              },
              "description": "List of currently active financial profile IDs",
              "title": "Profile Ids"
            },
            "description": "List of currently active financial profile IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountList"
                },
                "example": {
                  "accounts": [
                    {
                      "id": 1,
                      "user_id": 42,
                      "name": "Main Checking",
                      "initial_balance": 1000.0,
                      "current_balance": 1250.5,
                      "currency": "EUR",
                      "created_at": "2025-01-15T10:30:00Z",
                      "updated_at": "2025-01-20T14:22:00Z"
                    }
                  ],
                  "total": 1
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unauthorized profile IDs"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create new account",
        "description": "Create a new financial account for the authenticated user",
        "operationId": "create_account_api_v1_accounts__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get account by ID",
        "description": "Retrieve a specific account by its ID",
        "operationId": "get_account_api_v1_accounts__account_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          },
          "403": {
            "description": "Not authorized to access this account"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accounts"
        ],
        "summary": "Update account",
        "description": "Update an existing account (partial update supported)",
        "operationId": "update_account_api_v1_accounts__account_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Delete account",
        "description": "Delete an account and all its transactions (CASCADE)",
        "operationId": "delete_account_api_v1_accounts__account_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}/balance": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get account balance",
        "description": "Get current balance for a specific account with metadata",
        "operationId": "get_account_balance_api_v1_accounts__account_id__balance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalance"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountBalance": {
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "title": "Accountid",
            "description": "Account ID"
          },
          "balance": {
            "type": "string",
            "title": "Balance",
            "description": "Current account balance"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Account currency"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time",
            "title": "Lastupdated",
            "description": "Last transaction timestamp"
          }
        },
        "type": "object",
        "required": [
          "accountId",
          "balance",
          "currency",
          "lastUpdated"
        ],
        "title": "AccountBalance",
        "description": "Schema for account balance endpoint response.\nReturns current balance with metadata.",
        "example": {
          "account_id": "550e8400-e29b-41d4-a716-446655440050",
          "balance": 1250.5,
          "currency": "EUR",
          "last_updated": "2025-01-20T14:22:00Z"
        }
      },
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Account name",
            "examples": [
              "Main Checking Account",
              "Savings",
              "Credit Card"
            ]
          },
          "accountType": {
            "$ref": "#/components/schemas/AccountType",
            "description": "Type of account (checking, savings, credit_card, etc.)",
            "default": "checking"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "description": "ISO 4217 currency code (3 uppercase letters)",
            "default": "EUR",
            "examples": [
              "EUR",
              "USD",
              "GBP"
            ]
          },
          "institutionName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutionname",
            "description": "Name of the financial institution"
          },
          "creditLimit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creditlimit",
            "description": "Credit limit (for credit cards)"
          },
          "interestRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interestrate",
            "description": "Annual interest rate percentage (for loans/savings)"
          },
          "isIncludedInTotals": {
            "type": "boolean",
            "title": "Isincludedintotals",
            "description": "Include this account in net worth calculations",
            "default": true
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the account"
          },
          "financialProfileId": {
            "type": "string",
            "format": "uuid",
            "title": "Financialprofileid",
            "description": "ID of the financial profile this account belongs to (that is current user's default profile)"
          },
          "initialBalance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Initialbalance",
            "description": "Initial account balance (can be negative for debts)",
            "default": "0.00",
            "examples": [
              0,
              1000.5,
              -5000,
              5000
            ]
          },
          "accountNumberLast4": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4,
                "minLength": 4,
                "pattern": "^[0-9]{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountnumberlast4",
            "description": "Last 4 digits of account number"
          },
          "iban": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 34
              },
              {
                "type": "null"
              }
            ],
            "title": "Iban",
            "description": "Full IBAN (will be encrypted for high-security profiles)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "financialProfileId"
        ],
        "title": "AccountCreate",
        "description": "Schema for creating a new account.\nfinancial_profile_id is optional - if not provided, the user's default profile will be used.",
        "example": {
          "account_type": "savings",
          "currency": "EUR",
          "financial_profile_id": "550e8400-e29b-41d4-a716-446655440000",
          "initial_balance": 1000.0,
          "institution_name": "My Bank",
          "name": "Savings Account",
          "notes": "Primary savings account"
        }
      },
      "AccountList": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            },
            "type": "array",
            "title": "Accounts",
            "description": "List of accounts"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of accounts"
          }
        },
        "type": "object",
        "required": [
          "accounts",
          "total"
        ],
        "title": "AccountList",
        "description": "Schema for list accounts response with pagination.",
        "example": {
          "accounts": [
            {
              "account_type": "checking",
              "created_at": "2025-01-15T10:30:00Z",
              "currency": "EUR",
              "current_balance": 1250.5,
              "financial_profile_id": "550e8400-e29b-41d4-a716-446655440000",
              "id": "550e8400-e29b-41d4-a716-446655440050",
              "initial_balance": 1000.0,
              "institution_name": "My Bank",
              "is_active": true,
              "is_included_in_totals": true,
              "name": "Main Checking",
              "updated_at": "2025-01-20T14:22:00Z"
            }
          ],
          "total": 1
        }
      },
      "AccountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Account name",
            "examples": [
              "Main Checking Account",
              "Savings",
              "Credit Card"
            ]
          },
          "accountType": {
            "$ref": "#/components/schemas/AccountType",
            "description": "Type of account (checking, savings, credit_card, etc.)",
            "default": "checking"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "description": "ISO 4217 currency code (3 uppercase letters)",
            "default": "EUR",
            "examples": [
              "EUR",
              "USD",
              "GBP"
            ]
          },
          "institutionName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutionname",
            "description": "Name of the financial institution"
          },
          "creditLimit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creditlimit",
            "description": "Credit limit (for credit cards)"
          },
          "interestRate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interestrate",
            "description": "Annual interest rate percentage (for loans/savings)"
          },
          "isIncludedInTotals": {
            "type": "boolean",
            "title": "Isincludedintotals",
            "description": "Include this account in net worth calculations",
            "default": true
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the account"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique account identifier"
          },
          "financialProfileId": {
            "type": "string",
            "format": "uuid",
            "title": "Financialprofileid",
            "description": "ID of the financial profile this account belongs to"
          },
          "initialBalance": {
            "type": "string",
            "title": "Initialbalance",
            "description": "Initial balance when account was created (can be negative)"
          },
          "currentBalance": {
            "type": "string",
            "title": "Currentbalance",
            "description": "Current balance (initial + transactions)"
          },
          "accountNumberLast4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountnumberlast4",
            "description": "Last 4 digits of account number"
          },
          "iban": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iban",
            "description": "Full IBAN (encrypted for high-security profiles)"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "description": "Whether the account is currently active",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "Account creation timestamp (UTC)"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat",
            "description": "Last update timestamp (UTC)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "financialProfileId",
          "initialBalance",
          "currentBalance",
          "createdAt",
          "updatedAt"
        ],
        "title": "AccountResponse",
        "description": "Complete account schema returned by API endpoints.\nIncludes all fields including computed current_balance.",
        "example": {
          "account_type": "checking",
          "created_at": "2025-01-15T10:30:00Z",
          "currency": "EUR",
          "current_balance": 1250.5,
          "financial_profile_id": "550e8400-e29b-41d4-a716-446655440000",
          "id": "550e8400-e29b-41d4-a716-446655440050",
          "initial_balance": 1000.0,
          "institution_name": "My Bank",
          "is_active": true,
          "is_included_in_totals": true,
          "name": "Main Checking Account",
          "notes": "Primary checking account",
          "updated_at": "2025-01-20T14:22:00Z"
        }
      },
...
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Updated account name"
          },
          "accountType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated account type"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency",
            "description": "Updated currency code"
          },
          "initialBalance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialbalance",
            "description": "Initial account balance (can be negative for debts)",
            "default": "0.00",
            "examples": [
              0,
              1000.5,
              -5000,
              5000
            ]
          },
          "institutionName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutionname",
            "description": "Updated institution name"
          },
          "creditLimit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creditlimit",
            "description": "Updated credit limit"
          },
          "interestRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interestrate",
            "description": "Updated interest rate"
          },
          "isIncludedInTotals": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isincludedintotals",
            "description": "Updated inclusion in totals"
          },
          "accountNumberLast4": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4,
                "minLength": 4,
                "pattern": "^[0-9]{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountnumberlast4",
            "description": "Updated last 4 digits of account number"
          },
          "iban": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 34
              },
              {
                "type": "null"
              }
            ],
            "title": "Iban",
            "description": "Updated IBAN"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Updated notes"
          },
          "isActive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isactive",
            "description": "Whether the account is active"
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Schema for updating an existing account.\nAll fields are optional (partial update).",
        "example": {
          "credit_limit": 5000.0,
          "institution_name": "New Bank",
          "name": "Updated Account Name"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login endpoint. Token expires after 30 minutes."
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}